pyproject.toml
[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.poetry]
name = "gert"
version = "0.1.0"
description = "Ghent University Examination Results Tool"
authors = ["stijndcl"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
Pillow = "^9.4.0"
pypdf = "^3.3.0"
pdf2image = "^1.16.2"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
flake8-bugbear = "^23.1.20"
flake8-isort = "^6.0.0"
flake8-simplify = "^0.19.3"
pre-commit = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
